+++ audyn/utils/data/msd_recommend/__init__.py
@@ -1,0 +1,5 @@
+from ._download import download_interactions
+
+__all__ = [
+    "download_interactions",
+]
