+++ audyn/utils/data/msd_tagging/__init__.py
@@ -1,9 +1,9 @@
-from ._download import download_interactions, download_tags
+from ._download import download_tags
 
 __all__ = [
     "tags",
     "num_tags",
-    "download_interactions",
+    "download_tags",
 ]
 
 tags = download_tags()
